added some more player information
